backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "austrai-privacyproxy"
version = "2.0.0"
description = "Open-Source Anonymisierungsproxy für LLM-Anfragen. Erkennt sensible Daten, anonymisiert sie automatisch und re-hydriert die Antwort."
readme = "../README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Florian Bieche", email = "florian@flb.co.at"}]
keywords = ["privacy", "llm", "anonymization", "pii", "presidio", "gdpr", "dsgvo"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Text Processing",
]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "spacy>=3.8",
    "litellm>=1.70",
    "pydantic>=2.11",
    "pydantic-settings>=2.9",
    "python-multipart>=0.0.20",
    "sentence-transformers>=4.0",
    "PyMuPDF>=1.25",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    "Pillow>=11.0",
    "pytesseract>=0.3",
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
local-llm = ["llama-cpp-python>=0.3"]
dev = ["pytest>=8.0"]

[project.scripts]
privacyproxy = "app.cli:main"

[project.urls]
Homepage = "https://austr.ai"
Repository = "https://github.com/flbcoat/austrai-privacyproxy"
